requirements.txt
fastapi
uvicorn
pydantic
PyMuPDF
jinja2
langchain
langchain-google-vertexai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
